% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_login.R
\name{mod_login}
\alias{mod_login}
\title{mod_csv_file server function}
\usage{
mod_login(input, output, session, tabsPanel)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{stringsAsFactors}{logical: should character vectors be converted to factors?}
}
\description{
mod_csv_file server function
}
