% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_login.R
\name{mod_loginUI}
\alias{mod_loginUI}
\title{mod_csv_fileInput and mod_csv_file}
\usage{
mod_loginUI(id, label = "Log In")
}
\arguments{
\item{id}{shiny id}

\item{label}{fileInput label}
}
\description{
A shiny Module that imports csv file
}
\examples{
library(shiny)
library(DT)
if (interactive()){
ui <- fluidPage(
  mod_csv_fileInput("fichier"),
DTOutput("tableau")
)

server <- function(input, output, session) {
  data <- callModule(mod_csv_file,"fichier")
  output$tableau <- renderDT({data()})
}

shinyApp(ui, server)
}

}
