% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_login.R, R/mod_sample_browse.R,
%   R/mod_sample_register.R
\name{mod_login}
\alias{mod_login}
\alias{mod_sample_browse}
\alias{mod_sample_register}
\title{mod_login server function}
\usage{
mod_login(input, output, session)

mod_sample_browse(input, output, session)

mod_sample_register(input, output, session)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
mod_login server function

mod_login server function

mod_login server function
}
