% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_login.R
\name{mod_loginUI}
\alias{mod_loginUI}
\title{mod_loginUI and mod_login}
\usage{
mod_loginUI(id, label = "Log In")
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny Module for entering credentials and login into Airtable
}
\examples{
  mod_loginUI("login_body"))
  callModule(mod_login,"login_body")
}
